Align ID to Text
